EAS: add choosing pen color workflow
User can now choose a spesific color to paint a tile with when rainbow
mode is not on. Furthermore, the way a tile is painted is no longer
controlled by the `tile--painted` class. This way reseting tiles
painted by both random colors or by a chosen color can be done with the
tile.style.backgroundColor property